pyproject.toml
[project]
name = "postpuppy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.11",
    "prisma>=0.15.0",
    "python-dotenv>=1.0.1",
    "slack-bolt>=1.22.0",
    "starlette>=0.45.2",
    "uvicorn>=0.34.0",
    "uvloop>=0.21.0",
]

[project.scripts]
postpuppy = "postpuppy.__main__:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
]
